Conflicts:
	didactic-palm-tree/didactic-palm-tree.csproj

+++ didactic-palm-tree/Simulation/ITerminal.cs
@@ -2,7 +2,7 @@
 {
     public interface ITerminal
     {
-        int Voltage { get; set; }
+        double Voltage { get; set; }
         IConnection GetConnection();
         IComponent GetComponent();
     }
